Integrate completion into debugger
